Feat: Add Config-Driven Parameters
# Description
`appsettings.json` is added to the root of the project. It now stores
logging and DID configurations. This can be easily enriched in the
future.

Changes made significantly improves scalability and makes it easier to
twerk these configurations in the future if necessary.

As a result of some of these changes `LogConfig` became obsolete and is
now removed from the project.