Feat: Add JWT Based VCs (#31)
# Description
The `JWT` creation and verification is implemented.

- New service is added to the `Domain` layer to create and verify the
tokens
- New, token related, methods are added to the `Issuer` and `Verifier`
services
- Method to start the new flow is added to the `Presentation` 

This means that `VCs` are now wrapped inside `JWT` when created and
sent. This is more in accordance to the OpenID4VC specification.

Before, the `VCs` where sent just as they are between the entities when
running the prototype. This does not adhere to the OpenId4VC
specification.